+++ app/src/main/java/com/woowa/banchan/ui/common/worker/OrderWorker.kt
@@ -1,11 +1,11 @@
-package com.woowa.banchan.service
+package com.woowa.banchan.ui.common.worker
 
 import android.content.Context
 import androidx.hilt.work.HiltWorker
 import androidx.work.CoroutineWorker
 import androidx.work.WorkerParameters
 import com.woowa.banchan.domain.usecase.order.inter.UpdateOrderUseCase
-import com.woowa.banchan.ui.common.key.orderWorkerId
+import com.woowa.banchan.ui.common.key.ORDER_WORKER_ID
 import dagger.assisted.Assisted
 import dagger.assisted.AssistedInject
 
@@ -18,7 +18,7 @@ class OrderWorker @AssistedInject constructor(
 
     override suspend fun doWork(): Result {
 
-        updateOrderUseCase(inputData.getLong(orderWorkerId, 0), true)
+        updateOrderUseCase(inputData.getLong(ORDER_WORKER_ID, 0), true)
             .onSuccess { return Result.success() }
             .onFailure { return Result.failure() }
