+++ app/src/main/java/com/woowa/banchan/ui/home/HomeBaseFragment.kt
@@ -12,8 +12,8 @@ import androidx.fragment.app.Fragment
 import com.woowa.banchan.R
 import com.woowa.banchan.ui.common.bottomsheet.CartAddFragment
 import com.woowa.banchan.ui.common.event.EventObserver
-import com.woowa.banchan.ui.common.key.foodDetailHash
-import com.woowa.banchan.ui.common.key.foodDetailTitle
+import com.woowa.banchan.ui.common.key.FOOD_DETAIL_HASH
+import com.woowa.banchan.ui.common.key.FOOD_DETAIL_TITLE
 import com.woowa.banchan.ui.detail.DetailActivity
 import com.woowa.banchan.ui.home.adapter.HomeRVAdapter
 
@@ -62,8 +62,8 @@ abstract class HomeBaseFragment<T : ViewDataBinding>(@LayoutRes val layoutRes: I
             val intent = Intent(context, DetailActivity::class.java)
             val titleHash = it.split(",")
 
-            intent.putExtra(foodDetailTitle, titleHash[0])
-            intent.putExtra(foodDetailHash, titleHash[1])
+            intent.putExtra(FOOD_DETAIL_TITLE, titleHash[0])
+            intent.putExtra(FOOD_DETAIL_HASH, titleHash[1])
             startActivity(intent)
         })
     }
