+++ app/src/main/java/com/woowa/banchan/ui/order/OrderActivity.kt
@@ -10,7 +10,7 @@ import androidx.lifecycle.lifecycleScope
 import com.woowa.banchan.R
 import com.woowa.banchan.databinding.ActivityOrderBinding
 import com.woowa.banchan.ui.common.event.EventObserver
-import com.woowa.banchan.ui.common.key.orderId
+import com.woowa.banchan.ui.common.key.ORDER_ID
 import com.woowa.banchan.ui.common.uistate.UiState
 import com.woowa.banchan.ui.order.adapter.OrderRVAdapter
 import com.woowa.banchan.ui.order.detail.OrderDetailActivity
@@ -58,7 +58,7 @@ class OrderActivity : AppCompatActivity() {
 
         viewModel.orderItemClickEvent.observe(this, EventObserver {
             val intent = Intent(this, OrderDetailActivity::class.java)
-            intent.putExtra(orderId, it)
+            intent.putExtra(ORDER_ID, it)
             startActivity(intent)
         })
     }
